+++ CHANGELOG.md
@@ -4,7 +4,8 @@ First version containing base of the app with future updates in mind
 # V2
 Second version after feedback
 
-## Changes
+## C
+- Upd
 - Fix: changed native input to shadcn input in Service Card.
 - Update: add not established to service request client preferences, timeframes and time
 - Fix: Include only esitmated price into service card.
@@ -13,4 +14,4 @@ Second version after feedback
 - Feature: set date and time of service request creation
 - Update: add shadcn calendar inputs to every date time input, and enable manual date input
 - Update: change the order of blocks in service request
-- 
+- Update: Naprawy view, add ordering, fresh view, configuration for statuses, saving state between views, more columns to display, more actions, and context window with actions.